Extract key exchange tests
